# gh-130167: Minor `textwrap.dedent()` optimization