--- a/Lib/textwrap.py
+++ b/Lib/textwrap.py
@@ -432,26 +432,28 @@ def dedent(text):
     lines = text.split('\n')
 
     # Get length of leading whitespace, inspired by ``os.path.commonprefix()``.
-    l1 = None
-    l2 = None
+    val = False
     for i, line in enumerate(lines):
         # Compute min + max concurrently + normalize others
         if line and not line.isspace():
-            if l1 is None or line < l1:
-                l1 = line
-            if l2 is None or line > l2:
-                l2 = line
+            if val:
+                if line < l1:
+                    l1 = line
+                elif line > l2:
+                    l2 = line
+            else:
+               val = True
+               l1 = l2 = line 
+                
         else:
             lines[i] = ''
     
-    if l1 is None:
-        l1 = ''
+    if not val or not l1:
+        return '\n'.join(lines)
     
     for margin, c in enumerate(l1):
         if c != l2[margin] or c not in ' \t':
             break
-    else:
-        return '\n'.join(lines)
 
     return '\n'.join([line[margin:] for line in lines])